+++ src/components/Sparkbar/index.ts
@@ -1,2 +1,0 @@
-export {Sparkbar} from './Sparkbar';
-export type {SparkbarProps} from './Sparkbar';